# backend/api/openapi-bundled.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Multi Outlet POS Kopi Kopi & Inventory API",
    "version": "1.0.0",
    "description": "API Specification for Multi Outlet POS Kopi Popi & Inventory System."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local Development Server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "summary": "Register Customer",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Gilang"
                  },
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "rahasia123"
                  },
                  "confirm_password": {
                    "type": "string",
                    "example": "rahasia123"
                  },
                  "phone": {
                    "type": "string",
                    "example": "08123456789"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "example": "Gilang"
                        },
                        "email": {
                          "type": "string",
                          "example": "gilang@gmail.com"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/verify-email": {
      "post": {
        "summary": "Verify Email OTP",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  },
                  "otp_code": {
                    "type": "string",
                    "example": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "string",
                      "example": "email verified successfully"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Login",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "rahasia123"
                  },
                  "remember_me": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "token": {
                          "type": "string",
                          "example": "jwt_token"
                        },
                        "user": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
...
    "/auth/forgot-password": {
      "post": {
        "summary": "Forgot Password",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "gilang@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/auth/reset-password": {
      "post": {
        "summary": "Reset Password",
        "tags": [
          "Auth"
        ],
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "example": "123456"
                  },
                  "new_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  },
                  "confirm_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
...
    "/users/search": {
      "get": {
        "summary": "Search Customer by Name or Phone",
        "description": "Find a customer by name or phone number (for offline points redemption).",
        "tags": [
          "Employees"
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Phone number or customer name to search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/User"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get My Profile",
        "tags": [
          "Profile"
        ],
        "responses": {
          "200": {
            "description": "Success Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/User"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Profile",
        "description": "Partial update for user profile. PATCH is used because it updates only specific fields without replacing the entire object.",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Gilang Baru"
                  },
                  "phone": {
                    "type": "string",
                    "example": "081111111"
                  },
                  "profile_picture": {
                    "type": "string",
                    "example": "profile-1.jpg"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/me/password": {
      "patch": {
        "summary": "Update Password",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "current_password": {
                    "type": "string",
                    "example": "passwordlama"
                  },
                  "new_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  },
                  "confirm_password": {
                    "type": "string",
                    "example": "passwordbaru"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
...
    "/users/me/request-email-otp": {
      "post": {
        "summary": "Request Email Change OTP",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "new_email": {
                    "type": "string",
                    "example": "gilang_baru@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/me/email": {
      "put": {
        "summary": "Verify OTP & Update Email",
        "tags": [
          "Profile"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "otp": {
                    "type": "string",
                    "example": "123456"
                  },
                  "new_email": {
                    "type": "string",
                    "example": "gilang_baru@gmail.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/GenericError"
          }
        }
      }
    },
    "/users/managers": {
      "post": {
        "summary": "Create Manager (Admin Only)",
        "tags": [
          "Users (Employee Management)"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "branch_id": {
                    "type": "integer",
                    "example": 1
                  },
                  "name": {
                    "type": "string",
                    "example": "Budi Manager"
                  },
                  "email": {
                    "type": "string",
                    "example": "manager@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/cashiers": {
      "post": {
        "summary": "Create Cashier (Manager Only)",
        "tags": [
          "Users (Employee Management)"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Siti Kasir"
                  },
                  "email": {
                    "type": "string",
                    "example": "kasir@example.com"
                  },
                  "password": {
                    "type": "string",
                    "example": "password123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/users/{id}/disable": {
      "patch": {
        "summary": "Disable/Deactivate User (Admin Only)",
        "tags": [
          "Users (Employee Management)"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/categories": {
      "get": {
        "summary": "Get All Categories",
        "tags": [
          "Catalogues"
        ],
        "security": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Category"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Category (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "summary": "Update Category (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Delete Category (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/materials": {
      "get": {
        "summary": "Get All Materials (Admin/Manager Only)",
        "tags": [
          "Catalogues"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Material"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Material (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/materials/{id}": {
      "put": {
        "summary": "Update Material (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Delete Material (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/products": {
      "get": {
        "summary": "Get All Products",
        "description": "Public users see only basic info. Does NOT include BOM/Recipe.",
        "tags": [
          "Catalogues"
        ],
        "security": [],
        "parameters": [
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Product with Recipe (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "summary": "Get Product Detail",
        "description": "Includes Recipe (BOM) only if requester is Admin or Manager and adds ?include_recipe=true.",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "include_recipe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Product"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update Product and Recipe (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      },
      "delete": {
        "summary": "Soft Delete Product (Admin Only)",
        "tags": [
          "Catalogues"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
...
    "/branches/{id}/operating-hours": {
      "patch": {
        "summary": "Update Operating Hours (Admin Only)",
        "tags": [
          "Branches"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOperatingHoursRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "phone": {
            "type": "string"
          },
          "role": {
            "type": "string"
          },
          "profile_picture": {
            "type": "string"
          },
          "loyalty_points": {
            "type": "integer"
          }
        }
      },
...
      "Category": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string",
            "example": "Kopi"
          }
        }
      },
...
    },
    "responses": {
      "GenericError": {
        "description": "Error response",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
    }
  }
}